Applied code conventions and added explanations based on code in notebook
